+++ src/codemagic/tools/_app_store_connect/arguments.py
@@ -147,6 +147,15 @@ class Types:
         argument_type = bool
         environment_variable_key = "APP_STORE_CONNECT_ALTOOL_VERBOSE_LOGGING"
 
+    class MaxFindBuildWait(cli.TypedCliArgument[int]):
+        argument_type = int
+        environment_variable_key = "APP_STORE_CONNECT_MAX_FIND_BUILD_WAIT"
+        default_value = 10
+
+        @classmethod
+        def _is_valid(cls, value: int) -> bool:
+            return value > 0
+
     class MaxBuildProcessingWait(cli.TypedCliArgument[int]):
         argument_type = int
         environment_variable_key = "APP_STORE_CONNECT_MAX_BUILD_PROCESSING_WAIT"
@@ -894,6 +903,22 @@ class PublishArgument(cli.Argument):
             "Release notes either for TestFlight or App Store review submission. "
             "Describe what's new in this version of your app, "
             "such as new features, improvements, and bug fixes."
+        ),
+        argparse_kwargs={
+            "required": False,
+        },
+    )
+    MAX_BUILD_FIND_WAIT = cli.ArgumentProperties(
+        key="max_find_build_wait",
+        flags=("--max-find-build-wait",),
+        type=Types.MaxFindBuildWait,
+        description=(
+            "The maximum amount of minutes to wait for the freshly uploaded build "
+            "to become discoverable in App Store Connect. Works in conjunction with "
+            "TestFlight beta review submission or App Store review submission and "
+            "operations that depend on either one of those. If the build does not "
+            "become available within the specified timeframe, further submission "
+            "will be terminated."
         ),
         argparse_kwargs={
             "required": False,